Refactor filename extension handling in DynLibManager
The code that adds the correct extension for a dynamic library is
unsightly, broken (it tries to substr() a string without checking its
length; also, it doesn't build on Not-Windows) and duplicated.

This commit refactors this bit to address those issues.

Theoretical behavior modification: now, only DynLibManager tries to add
an extension to form a valid filename; DynLib assumes the given string
is a path.  In practice, only DynLibManager uses DynLib, so we're good,
I guess.